Starter code of conduct for the project (#21)
Set up a starter code of conduct for the project. This will later be moved to the administrative repository cardboardci/readme.

At the moment I just want to slap a code of conduct onto this repository (that can later be rolled out to the other images) to create a set of expectations.